# compose_api/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: compose-api
  version: 0.7.1
paths:
...
  /simulation/run:
    post:
      tags:
      - Simulation
      summary: Run a simulation
      description: '`simulator` names an owner-published image this deployment lists
        (settings.prebuilt_simulators);

        the job then runs in that image instead of the shared container. Omitted:
        the shared container.'
      operationId: run-simulation
      parameters:
      - name: interval_time
        in: query
        required: false
        schema:
          type: number
          default: 1.0
          title: Interval Time
      - name: batch_submission
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Batch Submission
      - name: simulator
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Simulator
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_run-simulation'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationExperiment'
        '400':
          description: The submission cannot run. For a problem with the request itself
            (empty file, unsupported type, interval out of range) `detail` is a string.
            For a problem with its contents (unreadable archive, or processes the
            registry does not allow) `detail` is an object with a `message` and a
            `violations` list naming each address.
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /simulations:
    get:
      tags:
      - Simulations
      summary: List simulations, newest first, each with its latest SLURM job
      operationId: list-simulations
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - submitting
            - waiting
            - queued
            - pending
            - running
            - completed
            - failed
            - cancelled
            - out_of_memory
            - suspended
            - timeout
            - unknown
            type: string
          - type: 'null'
          description: Only simulations in this state
          title: Status
        description: Only simulations in this state
      - name: simulator
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Only this prebuilt simulator (by name), or container definitions
            with this hash prefix
          title: Simulator
        description: Only this prebuilt simulator (by name), or container definitions
          with this hash prefix
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Only simulations created at or after this time
          title: Since
        description: Only simulations created at or after this time
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /simulations/{simulation_id}:
    get:
      tags:
      - Simulations
      summary: 'Describe one simulation: its latest SLURM job, and how many events
        and datasets its run recorded'
      operationId: get-simulation
      parameters:
      - name: simulation_id
        in: path
        required: true
        schema:
          type: integer
          title: Simulation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationDetail'
        '404':
          description: The simulation does not exist, or the caller may not read it
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_run-simulation:
      properties:
        uploaded_file:
          type: string
          title: Uploaded File
          format: binary
      type: object
      required:
      - uploaded_file
      title: Body_run-simulation
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SimulationDetail:
      properties:
        simulation_id:
          type: integer
          title: Simulation Id
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        experiment_id:
          type: string
          title: Experiment Id
        simulator_id:
          type: integer
          title: Simulator Id
        simulator:
          anyOf:
          - type: string
          - type: 'null'
          title: Simulator
        container_def_hash:
          type: string
          title: Container Def Hash
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: public
        status:
          type: string
          title: Status
        slurm_job_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Slurm Job Id
        start_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
          - type: 'null'
          title: End Time
        exit_code:
          anyOf:
          - type: integer
          - type: 'null'
          title: Exit Code
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        event_count:
          type: integer
          title: Event Count
          default: 0
        dataset_count:
          type: integer
          title: Dataset Count
          default: 0
      type: object
      required:
      - simulation_id
      - experiment_id
      - simulator_id
      - container_def_hash
      - status
      title: SimulationDetail
      description: A simulation with what its run recorded.
    SimulationExperiment:
      properties:
        simulation_database_id:
          type: integer
          title: Simulation Database Id
        simulator_database_id:
          type: integer
          title: Simulator Database Id
        last_updated:
          type: string
          title: Last Updated
        metadata:
          additionalProperties:
            type: string
          type: object
          title: Metadata
      type: object
      required:
      - simulation_database_id
      - simulator_database_id
      title: SimulationExperiment
    SimulationPage:
      properties:
        simulations:
          items:
            $ref: '#/components/schemas/SimulationSummary'
          type: array
          title: Simulations
        total:
          type: integer
          title: Total
        next_offset:
          anyOf:
          - type: integer
          - type: 'null'
          title: Next Offset
      type: object
      required:
      - simulations
      - total
      title: SimulationPage
    SimulationSummary:
      properties:
        simulation_id:
          type: integer
          title: Simulation Id
        created_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Created At
        experiment_id:
          type: string
          title: Experiment Id
        simulator_id:
          type: integer
          title: Simulator Id
        simulator:
          anyOf:
          - type: string
          - type: 'null'
          title: Simulator
        container_def_hash:
          type: string
          title: Container Def Hash
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: public
        status:
          type: string
          title: Status
        slurm_job_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Slurm Job Id
        start_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
          - type: 'null'
          title: End Time
        exit_code:
          anyOf:
          - type: integer
          - type: 'null'
          title: Exit Code
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
      type: object
      required:
      - simulation_id
      - experiment_id
      - simulator_id
      - container_def_hash
      - status
      title: SimulationSummary
      description: 'One simulation as a listing shows it: what ran, where, and how
        it went (its latest SLURM job).'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Visibility:
      type: string
      enum:
      - public
      - private
      title: Visibility
      description: Who may read a simulation and everything it produced (plan-observability
        O7).
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional Auth0 access token for this API. Omit it to call anonymously.
security:
- {}
- BearerAuth: []